+++ src/scene_server/host_server/host_service/api_init.go
@@ -1,22 +1,22 @@
 /*
  * Tencent is pleased to support the open source community by making 蓝鲸 available.
  * Copyright (C) 2017-2018 THL A29 Limited, a Tencent company. All rights reserved.
- * Licensed under the MIT License (the "License"); you may not use this file except 
+ * Licensed under the MIT License (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
  * http://opensource.org/licenses/MIT
  * Unless required by applicable law or agreed to in writing, software distributed under
  * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language governing permissions and 
+ * either express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
- 
+
 package ccapi
 
 import (
 	_ "configcenter/src/scene_server/host_server/host_service/actions/delete"
 	_ "configcenter/src/scene_server/host_server/host_service/actions/detail"
 	_ "configcenter/src/scene_server/host_server/host_service/actions/hosts"
-	_ "configcenter/src/scene_server/host_server/host_service/actions/openapi"
+	_ "configcenter/src/scene_server/host_server/host_service/actions/phpapi"
 	_ "configcenter/src/scene_server/host_server/host_service/actions/search"
 	_ "configcenter/src/scene_server/host_server/host_service/actions/update"
 	_ "configcenter/src/scene_server/host_server/host_service/actions/userapi"
