+++ src/scene_server/host_server/host_service/logics/host_add.go
@@ -274,7 +274,7 @@ func EnterIP(req *restful.Request, ownerID string, appID, moduleID int, ip strin
 
 	isExist, err := IsExistPlat(req, ObjAddr, common.KvMap{common.BKCloudIDField: cloudID})
 	if nil != err {
-		return errors.New(langHandle.Language("host_search_fail")) // "查询主机信息失败")
+		return errors.New(langHandle.Languagef("plat_get_str_err", err.Error())) // "查询主机信息失败")
 	}
 	if !isExist {
 		return errors.New(langHandle.Language("plat_id_not_exist"))
